refactor: split lexer functions and fix tokenization flow
